refactor(web): split schedule editor helpers [OPE-382]
