+++ crates/opengoose-web/src/data/schedules/editor/form.rs
@@ -1,143 +1,22 @@
-use anyhow::{Context, Result};
 use opengoose_persistence::{OrchestrationRun, Schedule};
 use opengoose_teams::scheduler;
 use urlencoding::encode;
 
-use super::catalog::ScheduleCatalog;
-use super::selection::{NEW_SCHEDULE_KEY, Selection};
+use super::super::catalog::ScheduleCatalog;
+use super::options::build_team_options;
 use crate::data::utils::{preview, run_tone};
-use crate::data::views::{
-    MetaRow, Notice, ScheduleEditorView, ScheduleHistoryItem, ScheduleListItem, SchedulesPageView,
-    SelectOption,
-};
+use crate::data::views::{MetaRow, Notice, ScheduleEditorView, ScheduleHistoryItem};
 
-pub(super) struct ScheduleDraft {
-    pub(super) original_name: Option<String>,
-    pub(super) name: String,
-    pub(super) cron_expression: String,
-    pub(super) team_name: String,
-    pub(super) input: String,
-    pub(super) enabled: bool,
+pub(in crate::data::schedules) struct ScheduleDraft {
+    pub(in crate::data::schedules) original_name: Option<String>,
+    pub(in crate::data::schedules) name: String,
+    pub(in crate::data::schedules) cron_expression: String,
+    pub(in crate::data::schedules) team_name: String,
+    pub(in crate::data::schedules) input: String,
+    pub(in crate::data::schedules) enabled: bool,
 }
 
-pub(super) fn build_page(
-    catalog: &ScheduleCatalog,
-    selection: Selection,
-    detail_override: Option<ScheduleEditorView>,
-) -> Result<SchedulesPageView> {
-    let total = catalog.schedules.len();
-    let enabled = catalog
-        .schedules
-        .iter()
-        .filter(|schedule| schedule.enabled)
-        .count();
-    let active_name = match &selection {
-        Selection::Existing(name) => Some(name.as_str()),
-        Selection::New => None,
-    };
-
-    Ok(SchedulesPageView {
-        mode_label: if total == 0 {
-            "Ready for first schedule".into()
-        } else {
-            format!("{enabled} active of {total}")
-        },
-        mode_tone: if total == 0 {
-            "neutral"
-        } else if enabled > 0 {
-            "success"
-        } else {
-            "amber"
-        },
-        schedules: catalog
-            .schedules
-            .iter()
-            .map(|schedule| build_schedule_list_item(schedule, active_name))
-            .collect(),
-        selected: if let Some(detail) = detail_override {
-            detail
-        } else {
-            match selection {
-                Selection::Existing(name) => build_existing_detail(
-                    catalog
-                        .schedules
-                        .iter()
-                        .find(|schedule| schedule.name == name)
-                        .context("selected schedule missing")?,
-                    catalog,
-                    None,
-                ),
-                Selection::New => build_new_detail(catalog, None, None),
-            }
-        },
-        new_schedule_url: format!("/schedules?schedule={NEW_SCHEDULE_KEY}"),
-    })
-}
-
-pub(super) fn build_error_page(
-    catalog: &ScheduleCatalog,
-    draft: &ScheduleDraft,
-    message: impl Into<String>,
-) -> Result<SchedulesPageView> {
-    let selection = draft
-        .original_name
-        .as_ref()
-        .map(|name| Selection::Existing(name.clone()))
-        .unwrap_or(Selection::New);
-    let detail = build_draft_detail(
-        catalog,
-        draft,
-        Some(Notice {
-            text: message.into(),
-            tone: "danger",
-        }),
-    );
-    build_page(catalog, selection, Some(detail))
-}
-
-fn build_schedule_list_item(schedule: &Schedule, active_name: Option<&str>) -> ScheduleListItem {
-    let next_label = schedule
-        .next_run_at
-        .as_deref()
-        .map(|value| format!("Next {value}"))
-        .unwrap_or_else(|| {
-            if schedule.enabled {
-                "Next fire pending".into()
-            } else {
-                "Paused".into()
-            }
-        });
-    ScheduleListItem {
-        title: schedule.name.clone(),
-        subtitle: format!(
-            "{} · {}",
-            schedule.team_name,
-            if schedule.input.is_empty() {
-                "default input"
-            } else {
-                "custom input"
-            }
-        ),
-        preview: format!("{} · {}", schedule.cron_expression, next_label),
-        source_label: schedule
-            .last_run_at
-            .as_deref()
-            .map(|value| format!("Last {value}"))
-            .unwrap_or_else(|| "Never run".into()),
-        status_label: if schedule.enabled {
-            "Enabled".into()
-        } else {
-            "Paused".into()
-        },
-        status_tone: if schedule.enabled { "sage" } else { "neutral" },
-        page_url: format!("/schedules?schedule={}", encode(&schedule.name)),
-        active: active_name
-            .map(|name| name == schedule.name.as_str())
-            .unwrap_or(false),
-    }
-}
-
-fn build_existing_detail(
+pub(super) fn build_existing_detail(
     schedule: &Schedule,
     catalog: &ScheduleCatalog,
     notice: Option<Notice>,
@@ -206,7 +85,7 @@ fn build_existing_detail(
     }
 }
 
-fn build_new_detail(
+pub(super) fn build_new_detail(
     catalog: &ScheduleCatalog,
     draft: Option<&ScheduleDraft>,
     notice: Option<Notice>,
@@ -260,7 +139,7 @@ fn build_new_detail(
     }
 }
 
-fn build_draft_detail(
+pub(super) fn build_draft_detail(
     catalog: &ScheduleCatalog,
     draft: &ScheduleDraft,
     notice: Option<Notice>,
@@ -332,31 +211,6 @@ fn build_draft_detail(
     build_new_detail(catalog, Some(draft), notice)
 }
 
-fn build_team_options(
-    installed_teams: &[String],
-    selected_team: Option<&str>,
-) -> Vec<SelectOption> {
-    let mut names = installed_teams.to_vec();
-    if let Some(selected_team) = selected_team
-        && !selected_team.is_empty()
-        && !names.iter().any(|team| team == selected_team)
-    {
-        names.push(selected_team.to_string());
-        names.sort();
-    }
-
-    names
-        .into_iter()
-        .map(|team| SelectOption {
-            selected: selected_team
-                .map(|selected| selected == team.as_str())
-                .unwrap_or(false),
-            label: team.clone(),
-            value: team,
-        })
-        .collect()
-}
-
 fn build_history(schedule: &Schedule, runs: &[OrchestrationRun]) -> Vec<ScheduleHistoryItem> {
     let expected_input = effective_schedule_input(schedule);
     runs.iter()
@@ -381,36 +235,86 @@ fn effective_schedule_input(schedule: &Schedule) -> String {
     }
 }
 
-pub(super) fn normalize_input(input: String, max_bytes: usize) -> String {
-    if input.trim().is_empty() {
-        String::new()
-    } else {
-        truncate_to_byte_boundary(&input, max_bytes)
+#[cfg(test)]
+mod tests {
+    use opengoose_persistence::RunStatus;
+
+    use super::super::super::catalog::ScheduleCatalog;
+    use super::*;
+
+    #[test]
+    fn build_history_uses_effective_schedule_input_and_limits_matches() {
+        let schedule = sample_schedule();
+        let mut runs = (0..10)
+            .map(|index| sample_run(&format!("run-{index}"), "ops", "Scheduled run: nightly-ops"))
+            .collect::<Vec<_>>();
+        runs.push(sample_run(
+            "wrong-team",
+            "infra",
+            "Scheduled run: nightly-ops",
+        ));
+        runs.push(sample_run("wrong-input", "ops", "different input"));
+
+        let history = build_history(&schedule, &runs);
+
+        assert_eq!(history.len(), 8);
+        assert_eq!(history[0].title, "run-0");
+        assert_eq!(history[7].title, "run-7");
     }
-}
 
-pub(super) fn normalize_trimmed_field(value: &str, max_bytes: usize) -> String {
-    truncate_to_byte_boundary(value.trim(), max_bytes)
-}
+    #[test]
+    fn build_draft_detail_falls_back_to_new_detail_for_missing_original_schedule() {
+        let catalog = ScheduleCatalog {
+            schedules: vec![],
+            runs: vec![],
+            installed_teams: vec!["ops".into()],
+        };
+        let detail = build_draft_detail(
+            &catalog,
+            &ScheduleDraft {
+                original_name: Some("missing".into()),
+                name: "missing".into(),
+                cron_expression: "0 0 * * * *".into(),
+                team_name: "ops".into(),
+                input: String::new(),
+                enabled: true,
+            },
+            None,
+        );
 
-pub(super) fn normalize_optional_field(value: Option<&str>, max_bytes: usize) -> Option<String> {
-    value
-        .map(|item| normalize_trimmed_field(item, max_bytes))
-        .filter(|item| !item.is_empty())
-}
-
-pub(super) fn trimmed_len_exceeds(value: &str, max_bytes: usize) -> bool {
-    value.trim().len() > max_bytes
-}
-
-fn truncate_to_byte_boundary(value: &str, max_bytes: usize) -> String {
-    if value.len() <= max_bytes {
-        return value.to_owned();
+        assert!(detail.is_new);
+        assert_eq!(detail.title, "Create schedule");
+        assert_eq!(detail.team_name, "ops");
     }
 
-    let mut end = max_bytes;
-    while end > 0 && !value.is_char_boundary(end) {
-        end -= 1;
+    fn sample_schedule() -> Schedule {
+        Schedule {
+            id: 1,
+            name: "nightly-ops".into(),
+            cron_expression: "0 0 * * * *".into(),
+            team_name: "ops".into(),
+            input: String::new(),
+            enabled: true,
+            last_run_at: None,
+            next_run_at: Some("2026-03-12 00:00:00".into()),
+            created_at: "2026-03-11 00:00:00".into(),
+            updated_at: "2026-03-11 00:00:00".into(),
+        }
     }
-    value[..end].to_owned()
+
+    fn sample_run(id: &str, team_name: &str, input: &str) -> OrchestrationRun {
+        OrchestrationRun {
+            team_run_id: id.into(),
+            session_key: format!("session-{id}"),
+            team_name: team_name.into(),
+            workflow: "chain".into(),
+            input: input.into(),
+            status: RunStatus::Completed,
+            current_step: 1,
+            total_steps: 1,
+            result: None,
+            created_at: "2026-03-11 00:00:00".into(),
+            updated_at: "2026-03-11 00:00:00".into(),
+        }
+    }
 }
